+++ tests/cases/motion/cursor/run.vim
@@ -2,7 +2,7 @@ source $TIRENVI_ROOT/tests/common.vim
 
 lua << EOF
 function print_wrap(title)
-    local Context = require("tirenvi.app.context")
+    local Context = require("tirenvi.io.context")
 	ctx =  Context.from_buf()
     require("tirenvi.app.pipeline").auto_wrap(ctx)
     print(tostring(vim.wo[0].wrap) .. " : " .. (title or ""))
