+++ tests/cases/ut/cursor/run.vim
@@ -3,7 +3,7 @@ source $TIRENVI_ROOT/tests/common.vim
 new
 
 lua << EOF
-  Context = require("tirenvi.app.context")
+  Context = require("tirenvi.io.context")
   Buffer = require("tirenvi.io.buffer")
   Range = require("tirenvi.util.range")
   local lines = {
