+++ tests/cases/ut/log/cursor/run.vim
@@ -3,7 +3,7 @@ source $TIRENVI_ROOT/tests/common.vim
 new
 
 lua << EOF
-  Context = require("tirenvi.app.context")
+  Context = require("tirenvi.io.context")
   Buffer = require("tirenvi.io.buffer")
   Range = require("tirenvi.util.range")
   CursorNvim = require("tirenvi.cursor.nvim")
